fix: première correction.. ou pas

+++ backend/API-Ultimate/src/main/java/org/ultimateam/apiultimate/service/JoueurService.java
@@ -123,45 +123,45 @@ public class JoueurService {
         Joueur joueur = getById(id_joueur);
         Equipe equipe =equipeService.getById(id_equipe);
         if (equipe.getJoueurs().contains(joueur)) {
-            throw new ResponseStatusException(HttpStatus.FORBIDDEN,"Le joueur fait déjà parti de l'équipe");
+            throw new ResponseStatusException(HttpStatus.METHOD_NOT_ALLOWED,"Le joueur fait déjà parti de l'équipe");
         }
         else if (equipe.isFull())
-            throw new ResponseStatusException(HttpStatus.FORBIDDEN,"Equipe complete");
+            throw new ResponseStatusException(HttpStatus.METHOD_NOT_ALLOWED,"Equipe complete");
         else if (equipe.getGenre() == Genre.H2F3){
             if (joueur.getGenre() == GenreJoueur.HOMME && getNbHommes(equipe) >=2 ){
-                throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Nombre d'homme maximum atteint");
+                throw new ResponseStatusException(HttpStatus.METHOD_NOT_ALLOWED, "Nombre d'homme maximum atteint");
             }
             else if (joueur.getGenre() == GenreJoueur.FEMME && getNbFemmes(equipe) >=3 ){
-                throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Nombre de femme maximum atteint");
+                throw new ResponseStatusException(HttpStatus.METHOD_NOT_ALLOWED, "Nombre de femme maximum atteint");
             }
         }
         else if (equipe.getGenre() == Genre.H3F2){
             if (joueur.getGenre() == GenreJoueur.HOMME && getNbHommes(equipe) >=3 ){
-                throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Nombre d'homme maximum atteint");
+                throw new ResponseStatusException(HttpStatus.METHOD_NOT_ALLOWED, "Nombre d'homme maximum atteint");
             }
             else if (joueur.getGenre() == GenreJoueur.FEMME && getNbFemmes(equipe) >=2 ){
-                throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Nombre de femme maximum atteint");
+                throw new ResponseStatusException(HttpStatus.METHOD_NOT_ALLOWED, "Nombre de femme maximum atteint");
             }
         }
         else if (equipe.getGenre() == Genre.H4F3){
             if (getNbHommes(equipe) >=4 ){
-                throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Nombre d'homme maximum atteint");
+                throw new ResponseStatusException(HttpStatus.METHOD_NOT_ALLOWED, "Nombre d'homme maximum atteint");
             }
             else if (joueur.getGenre() == GenreJoueur.FEMME && getNbFemmes(equipe) >=3 ){
-                throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Nombre de femme maximum atteint");
+                throw new ResponseStatusException(HttpStatus.METHOD_NOT_ALLOWED, "Nombre de femme maximum atteint");
             }
         }
         else if (equipe.getGenre() == Genre.H3F4){
             if (joueur.getGenre() == GenreJoueur.HOMME && getNbHommes(equipe) >=3 ){
-                throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Nombre d'homme maximum atteint");
+                throw new ResponseStatusException(HttpStatus.METHOD_NOT_ALLOWED, "Nombre d'homme maximum atteint");
             }
             else if (joueur.getGenre() == GenreJoueur.FEMME && getNbFemmes(equipe) >=4 ){
-                throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Nombre de femme maximum atteint");
+                throw new ResponseStatusException(HttpStatus.METHOD_NOT_ALLOWED, "Nombre de femme maximum atteint");
             }
         }
         else if (equipe.getGenre() == Genre.FEMME){
             if (joueur.getGenre() == GenreJoueur.HOMME){
-                throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Equipe ne peut pas contenir d'hommes");
+                throw new ResponseStatusException(HttpStatus.METHOD_NOT_ALLOWED, "Equipe ne peut pas contenir d'hommes");
             }
         }
         equipe.addJoueur(joueur);
@@ -199,7 +199,7 @@ public class JoueurService {
                 .orElseThrow(() -> new EntityNotFoundException("Équipe non trouvée"));
 
         if (joueur.getEquipe() != null) {
-            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Impossible de rejoindre plusieurs équipes");
+            throw new ResponseStatusException(HttpStatus.METHOD_NOT_ALLOWED, "Impossible de rejoindre plusieurs équipes");
         }
         JoueurRequest request = new JoueurRequest(joueur, equipe);
         request.setId(request.getId());
@@ -220,15 +220,16 @@ public class JoueurService {
      */
     public Joueur accepterDemande(long idJoueur, long idEquipe) {
         JoueurRequestId id = new JoueurRequestId(idJoueur, idEquipe);
-        JoueurRequest joueurRequest = joueurRequestRepository.findById(id)
-                .orElseThrow(() -> new ResponseStatusException(
-                        HttpStatus.NOT_FOUND,
-                        "Demande inexistante"
-                ));
+        JoueurRequest joueurRequest = joueurRequestRepository.findById(id).orElse(null);
+        if (joueurRequest == null) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Demande inexistante");
+
+        }
+        Joueur joueur = assignerEquipe(joueurRequest.getId().getIdJoueur(), joueurRequest.getId().getIdEquipe());
         joueurRequestRepository.deleteById(id);
         List<JoueurRequest> toDelet = joueurRequestRepository.findAll().stream().filter(req -> Objects.equals(req.getId().getIdJoueur(), joueurRequest.getId().getIdJoueur())).toList();
         joueurRequestRepository.deleteAll(toDelet);
-        return assignerEquipe(joueurRequest.getId().getIdJoueur(), joueurRequest.getId().getIdEquipe());
+        return joueur;
     }
 
     /**
@@ -249,7 +250,6 @@ public class JoueurService {
     /**
      * Récupère les joueurs n'appartenant à aucune équipe (joueurs "solo").
      *
-     * @param genre filtre optionnel par genre (peut être {@code null})
      * @return la liste des {@link Joueur} solo correspondant au filtre
      */
     public List<Joueur> getJoueurSolo(Long idEquipe) {
